Html & Css
This is a simple project on HTML & CSS. In this HTML & CSS, you need to learn to do some stuff on the navigation bar, edit stuff on videos or images, login page/signup page, and animation objects which is basic of creating a good website. In index.html, index.css is the basic of this language you need to know. For more just keep practicing in HTML CSS.